Moved `events` module into `dynamic`
