+++ crates/wolf_engine_events/src/dynamic/mod.rs
@@ -3,6 +3,7 @@
 mod event;
 pub use event::*;
 
+pub mod events;
 pub mod event_loop;
 
 mod event_sender;
