+++ crates/wolf_engine_events/src/lib.rs
@@ -106,6 +106,5 @@
 mod event_queue;
 
 pub mod dynamic;
-pub mod events;
 pub use event_queue::*;
 pub mod mpsc;
